Refactor replace_ellipsis() out of runner.docker
This will soon be re-used by a second runner.
